refactor: 💡 move validate function rely on model
move validate function rely on model and fields of model

BREAKING CHANGE: 🧨 validators folder
