+++ src/validators/courseValidator.js
@@ -1,11 +1,72 @@
-import {
-  validateCourseName,
-  validateCourseCost,
-  validateCourseDescription,
-  validateCourseThumbnail,
-  validateCourseUserId,
-  returnValidationResult
-} from "./index.js";
+import { returnValidationResult } from "./index.js";
+import { body } from "express-validator";
+
+export const validateCourseName = async (req, isOptional = false) => {
+  await body("name")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .trim()
+    .notEmpty()
+    .withMessage("Course name is required")
+    .matches(`[A-Za-z]+`)
+    .withMessage("Course name must include A-Z")
+    .run(req);
+};
+
+export const validateCourseCost = async (req, isOptional = false) => {
+  await body("cost")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .notEmpty()
+    .withMessage(`Course cost is required`)
+    .matches(/^[1-9][0-9]*$/)
+    .withMessage(`Course cost must greater than 0`)
+    .run(req);
+};
+
+export const validateCourseDescription = async (req, isOptional = false) => {
+  await body("description")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .trim()
+    .notEmpty()
+    .withMessage("Course description is required")
+    .matches(`[A-Za-z]+`)
+    .withMessage("Course description must include A-Z")
+    .run(req);
+};
+
+export const validateCourseThumbnail = async (req, isOptional = false) => {
+  await body("thumbnail")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .trim()
+    .notEmpty()
+    .withMessage("Course thumbnail is required")
+    .matches(
+      /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_\+.~#?&//=]*)/
+    )
+    .withMessage("Course thumbnail is invalid")
+    .run(req);
+};
+
+export const validateCourseUserId = async (req) => {
+  await body("userId")
+    .trim()
+    .notEmpty()
+    .withMessage("Course userId is required")
+    .isLength({ min: 24, max: 24 })
+    .withMessage("Course userId must include 24 characters")
+    .run(req);
+};
 
 export const validateCourse = async (req) => {
   await Promise.all([
