+++ src/validators/lessonValidator.js
@@ -1,8 +1,35 @@
-import {
-  validateLessonName,
-  validateVideoId,
-  returnValidationResult
-} from "./index.js";
+import { returnValidationResult } from "./index.js";
+import { body } from "express-validator";
+
+export const validateLessonName = async (req, isOptional = false) => {
+  await body("name")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .trim()
+    .notEmpty()
+    .withMessage("Name lesson is required")
+    .matches(`[A-Za-z]+`)
+    .withMessage("Name lesson must include A-Z")
+    .isLength({ min: 10, max: 100 })
+    .withMessage("Name lesson must from 10 to 100 characters")
+    .run(req);
+};
+
+export const validateVideoId = async (req, isOptional = false) => {
+  await body("videoId")
+    .optional({
+      checkFalsy: isOptional,
+      nullable: isOptional
+    })
+    .trim()
+    .notEmpty()
+    .withMessage("Video Id is required")
+    .isLength(11)
+    .withMessage("Video Id is invalid")
+    .run(req);
+};
 
 export const validateLesson = async (req) => {
   await Promise.all([validateLessonName(req), validateVideoId(req)]);
